+++ Device/Memory/gpuMemManager.h
@@ -15,12 +15,12 @@
 class GPUMemManager
 {
 public:
-	static void MemcpyHostToDevice(void *pHostSrc, void *pDevDst, int numofByte);
-	static void MemcpyDeviceToHost(void *pDevSrc, void *pHostDst, int numofByte);
-	static void MemcpyDeviceToDevice(void *pDevSrc, void *pDevDst, int numofByte);
-	static void Memset(void *pDevSrc, int value, int numofByte);
+	static void MemcpyHostToDevice(void *pHostSrc, void *pDevDst, long long numofByte);
+	static void MemcpyDeviceToHost(void *pDevSrc, void *pHostDst, long long numofByte);
+	static void MemcpyDeviceToDevice(void *pDevSrc, void *pDevDst, long long numofByte);
+	static void Memset(void *pDevSrc, int value, long long numofByte);
 
-	void TestMemcpyHostToDevice(void *pHostSrc, void *pDevDst, int numofByte);
+	void TestMemcpyHostToDevice(void *pHostSrc, void *pDevDst, long long numofByte);
 	void TestMemcpyDeviceToHost();
 	void TestMemcpyDeviceToDevice();
 };
